% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_info.R
\name{clust_info}
\alias{clust_info}
\title{Wot?}
\usage{
clust_info(mi, score)
}
\arguments{
\item{mi}{output from \code{match_info}}

\item{score}{matching variable (1-dimensional)}

\item{output}{'cluster' for statistics on cluster level, 'delta'
for score differences on individual level and 'both' for both
(as list)}
}
\value{
data frame of info, or list of data frames
}
\description{
Extra cluster info
}
\details{
Some additional details I find useful in connection with a
    fullmatch (or similar matching)
}
\author{
Henrik Renlund
}

