% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-curv.R
\name{make_glist}
\alias{make_glist}
\title{helper function for 'surv_curv'}
\usage{
make_glist(x, ref = NULL)
}
\arguments{
\item{x}{thing to create glist from}

\item{ref}{reference for thing}
}
\description{
helper function for 'surv_curv'
}
