% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_info.R
\name{match_info}
\alias{match_info}
\title{individual level information}
\usage{
match_info(data, tr, cl, id = NULL, trv = 1, keep = NULL, warn = TRUE)
}
\arguments{
\item{data}{the data}

\item{tr}{the binary treatment variable, as character}

\item{cl}{the clustering variable, as character}

\item{id}{the id variable (optional), as character. If this is
used then only the relevant variables (+ those specified by 'keep') will be
returned (in the same order as data).}

\item{trv}{the value of the treatment (of variable \code{tr}), '1'
by default}

\item{keep}{character vector of the names of variables you'd like
to keep (id id is not  \code{NULL})}

\item{warn}{warn in case of emergency?}
}
\value{
data frame (\code{tbl_df}) with new parameters \code{tr_n} the number of
    treated within the cluster,  \code{ctrl_n} the number of
    controls within the cluster, \code{cl.weight} the weight,
    \code{cid} for describing the match it is useful to have a
    'cluster id', use this with e.g. \code{dplyr::group_by(tr, cid)} and
    \code{summarise} with functions using \code{weight =
    cl.weight} to get stats for weighted treated and control statistics.
}
\description{
Extra fullmatch info
}
\details{
Some additional details I find useful in connection with a
    fullmatch (or similar matching)
}
\examples{
df <- data_frame(
   id = 1:13,
   foo = c(0,0,1, 0,1, 1,1,0, 1,1,0,0, 0),
   bar = c(rep(c(letters[1:4]), c(3,2,3,4)), NA),
   x = round(runif(13),2)
)
mi <- match_info(data = df, tr = "foo", cl = "bar")
attributes(mi)
match_info(data = df, tr = "foo", cl = "bar", id = "id")
df$foo <- ifelse(df$foo == 1, "Treated", "Control")
match_info(data = df, tr = "foo", cl = "bar", trv = "Treated")
}
\author{
Henrik Renlund
}

