% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-curv.R
\name{surv_curv}
\alias{surv_curv}
\title{multiple survival curves}
\usage{
surv_curv(data, surv, strata = NULL, glist = NULL, w = NULL,
  prefix = c(event = "ev.", time = "t."), progress = FALSE,
  stringsAsFactors = TRUE)
}
\arguments{
\item{data}{the data}

\item{surv}{the \code{Surv} objects (as character, refering to
variables in 'data') OR the common names of time + event, if
these have a consistent prefix.}

\item{strata}{stratifying variable}

\item{glist}{a grouping list, or the name of a grouping variable
in 'data'}

\item{w}{weight or name of variable that provides weights}

\item{prefix}{prefix values for times and events}

\item{progress}{primitive displayer of progress}

\item{stringsAsFactors}{logical, if TRUE will keep ordering of ingoing stuff}
}
\value{
A \code{tbl_df} with an \code{rbind}ed \code{broom::tidy}'d
    \code{survival::survfit}ed version of all the combinations of
    'strata' and 'glist' for each 'surv'. That is a crap
    explanation...
}
\description{
create multiple survival curves for multiple endpoints,
    stratifications and groupings
}
\examples{
n = 1200
df <- data_frame(
        t.foo = abs(rnorm(n, 10, 2)),
        ev.foo = rbinom(n, 1, 0.2),
        t.bar = abs(rnorm(n, 10, 3)),
        ev.bar = rbinom(n, 1, 0.3),
        strutt = sample(letters[1:2], n, TRUE),
        gsak = sample(LETTERS[4:7], n, TRUE),
        wait = .5 + abs(rnorm(n, 0, 1))
)
surv_curv(data = df, surv = c("foo", "bar"), w = "wait",
          strata = "strutt", glist = "gsak")
}

